+++ frontend/src/util/assets.ts
